feat(db): refactor person repository into specialized components
This commit introduces a major architectural refactoring of the Person repository, decomposing the monolithic `PersonRepositoryImpl` into seven specialized, single-responsibility repositories. This change improves modularity, maintainability, and testability across the person domain.

Components affected:
- `banking-db-postgres`: Decomposed `PersonRepositoryImpl` into `PersonRepositoryImpl`, `CountryRepositoryImpl`, `CountrySubdivisionRepositoryImpl`, `LocalityRepositoryImpl`, `LocationRepositoryImpl`, `MessagingRepositoryImpl`, and `EntityReferenceRepositoryImpl`.
- `banking-db`: Updated `PersonRepository` trait to reflect the new, more granular structure.
- `banking-logic`: Updated `PersonServiceImpl` to use the new specialized repositories.
- `banking-api`: Minor updates to the `Person` domain model.

Benefits:
- Improved Modularity: Each new repository manages a specific aggregate, improving code clarity and reducing complexity.
- Enhanced Maintainability: Smaller, focused repositories are easier to understand, test, and maintain.
- Testability: The new structure allows for more granular and isolated testing, as evidenced by the 31+ new unit and integration tests.
- Scalability: This decomposition lays a solid foundation for future expansion of the Person domain.
