+++ .roo/commands/model-caches.md
@@ -12,7 +12,7 @@ description: you are a Rust programming expert tasked with generating a thread-s
 1.  **Parse Comment and Check for Existing Cache**:
     -   **Trigger:** The generation process is initiated by a `/// # Cache: <CacheName>` instruction in the comment block above a main model struct (e.g., `CountryModel`).
     -   **Idempotency Check:** Before generating any code, scan the target file for the existence of `pub struct <CacheName>`. If this struct and its corresponding `impl` block already exist, the script must halt execution for this model to ensure idempotency and prevent duplicate code.
-    -   **Source Index Model:** The cache is built for the corresponding `<ModelName>IdxModel` struct, which must already be defined or generated in the same file.
+    -   **Source Index Model:** The cache is built for the corresponding `<ModelName>IdxModel` struct, which must already be defined or generated in the same file. The `IdxModel` must contain `version: i32` and `hash: i64` fields.
     -   **Cache Properties:** Parse the attributes listed under the `# Cache` block:
         -   `- Concurent`: Indicates the cache must be thread-safe and wrapped in `Arc<Self>`.
         -   `- <Immutability>`: Defines the cache's behavior (e.g., `Immutable Set of Immutable Records`).
